+++ client/Cargo.toml
@@ -15,7 +15,6 @@ serde = { workspace = true, features = ["std"] }
 serde_json = { workspace = true }
 serde_with = { workspace = true }
 sp-core = { workspace = true, features = ["std"] }
-sp-runtime = { workspace = true, features = ["std"] }
 tokio = { workspace = true }
 base64 = { workspace = true }
 thiserror = { workspace = true }
