+++ client/src/provider_node_request_scheme/bucket.rs
@@ -1,0 +1,19 @@
+// SPDX-License-Identifier: GPL-3.0-only
+
+use serde::{Deserialize, Serialize};
+use storage_primitives::BucketId;
+
+/// Bucket summary info.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct BucketSummary {
+    pub bucket_id: BucketId,
+    pub mmr_root: String,
+    pub start_seq: u64,
+    pub leaf_count: u64,
+}
+
+/// Response with bucket list.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ListBucketsResponse {
+    pub buckets: Vec<BucketSummary>,
+}
