+++ client/src/provider_node_request_scheme/mod.rs
@@ -1,0 +1,27 @@
+// SPDX-License-Identifier: GPL-3.0-only
+
+//! API types for the provider node.
+
+pub mod agreement;
+pub mod bucket;
+pub mod commit;
+pub mod commitment;
+pub mod delete;
+pub mod existence;
+pub mod info;
+pub mod proof;
+pub mod read;
+pub mod replica_sync;
+pub mod upload_download;
+
+pub use agreement::*;
+pub use bucket::*;
+pub use commit::*;
+pub use commitment::*;
+pub use delete::*;
+pub use existence::*;
+pub use info::*;
+pub use proof::*;
+pub use read::*;
+pub use replica_sync::*;
+pub use upload_download::*;
