+++ provider-node/src/lib.rs
@@ -28,7 +28,6 @@ pub mod s3_api;
 pub mod s3_index;
 pub mod storage;
 pub(crate) mod subxt_client;
-pub mod types;
 
 pub use api::create_router;
 pub use chain_state_coordinator::{
@@ -46,7 +45,7 @@ pub use checkpoint_coordinator::{
 };
 pub use error::Error;
 pub use fs_index::FsIndexManager;
-pub use negotiate::{AgreementTermsOf, NegotiateRequest, NonceCounter, SignedTerms};
+pub use negotiate::NonceCounter;
 pub use replica_sync::ReplicaSync;
 pub use replica_sync_coordinator::{
     ReplicaSyncChainClient, ReplicaSyncCoordinator, ReplicaSyncCoordinatorConfig,
@@ -57,7 +56,7 @@ pub use storage::{
     build_merkle_proof, build_padded_merkle_tree, hex_decode, hex_encode, BucketInfo,
     DiskNonceStore, DiskStorage, NonceStore, NullNonceStore, Storage, StorageBackend, StoredNode,
 };
-pub use types::*;
+pub use storage_client::provider_node_request_scheme::*;
 
 use std::str::FromStr;
 use std::sync::Arc;
