+++ storage-interfaces/file-system/client/examples/basic_usage.rs
@@ -22,9 +22,9 @@
 //! ```
 
 use file_system_client::FileSystemClient;
-use sp_runtime::AccountId32;
 use std::env;
 use storage_client::{NegotiateRequest, ProviderClient};
+use storage_subxt::subxt::utils::AccountId32;
 use storage_subxt::subxt_signer::sr25519::dev as dev_signer;
 
 const DEFAULT_CHAIN_WS: &str = "ws://127.0.0.1:2222";
