+++ storage-interfaces/s3/client/Cargo.toml
@@ -14,7 +14,6 @@ storage-client = { workspace = true }
 
 # Substrate primitives
 sp-core = { workspace = true, features = ["std"] }
-sp-runtime = { workspace = true, features = ["std"] }
 
 # Async/HTTP
 reqwest = { workspace = true }
