+++ storage-interfaces/s3/client/examples/basic_usage.rs
@@ -5,10 +5,10 @@
 //! Usage: cargo run --example basic_usage [chain_ws] [provider_url] [seed]
 
 use s3_client::{PutObjectOptions, S3Client};
-use sp_runtime::AccountId32;
 use std::collections::HashMap;
 use std::env;
 use storage_client::{NegotiateRequest, ProviderClient};
+use storage_subxt::subxt::utils::AccountId32;
 use storage_subxt::subxt_signer::sr25519::dev as dev_signer;
 
 const DEFAULT_CHAIN_WS: &str = "ws://127.0.0.1:2222";
